+++ test/formatters/samplecode/nixfmt/out.nix
@@ -1,0 +1,15 @@
+{
+
+  nixpkgs = {
+    config.allowUnfree = true;
+    overlays = [
+      (final: prev: {
+        unstable = import self.inputs.nixpkgs-unstable {
+          system = system;
+          config.allowUnfree = true;
+        };
+      })
+    ];
+  };
+
+}