+++ apheleia.el
@@ -8,7 +8,7 @@
 ;; Keywords: tools
 ;; Package-Requires: ((emacs "26"))
 ;; SPDX-License-Identifier: MIT
-;; Version: 2.0
+;; Version: 3.0
 
 ;;; Commentary: