+++ CHANGELOG.md
@@ -14,13 +14,14 @@ The format is based on [Keep a Changelog].
 * Emacs 25 is no longer supported.
 
 ### Formatters
-* [dart-format](https://dart.dev/tools/dart-format) for Dart
+* [dart-format](https://dart.dev/tools/dart-format) for Dart ([#89]).
 
 ### Features
 * Support remote files and buffers that were opened through TRAMP
   ([#33]).
 
 [#33]: https://github.com/raxod502/apheleia/issues/33
+[#89]: https://github.com/raxod502/apheleia/pull/89
 
 ## 2.0
 ### Breaking changes