+++ CHANGELOG.md
@@ -32,6 +32,8 @@ The format is based on [Keep a Changelog].
 [#87]: https://github.com/raxod502/apheleia/pull/87
 [#89]: https://github.com/raxod502/apheleia/pull/89
 [#94]: https://github.com/radian-software/apheleia/pull/94
+[#95]: https://github.com/radian-software/apheleia/pull/95
+[#98]: https://github.com/radian-software/apheleia/pull/98
 
 ## 2.0
 ### Breaking changes